refactor: separate cart page components into cart-page/
Move 7 page-only cart components from components/cart/ to
components/cart-page/. components/cart/ now contains only global
infrastructure (context, actions, overlay).
