+++ apps/template/app/cart/page.tsx
@@ -4,12 +4,12 @@ import { getMessages, getTranslations } from "next-intl/server";
 import { Suspense } from "react";
 
 import { CartContextSync } from "@/components/cart/context-sync";
-import { Empty } from "@/components/cart/empty-cart";
-import { Header } from "@/components/cart/header";
-import { ItemsSection } from "@/components/cart/items-section";
-import { PageSkeleton } from "@/components/cart/skeletons";
-import { Summary } from "@/components/cart/summary";
-import { Upsells } from "@/components/cart/upsell-recommendations";
+import { Empty } from "@/components/cart-page/empty-cart";
+import { Header } from "@/components/cart-page/header";
+import { ItemsSection } from "@/components/cart-page/items-section";
+import { PageSkeleton } from "@/components/cart-page/skeletons";
+import { Summary } from "@/components/cart-page/summary";
+import { Upsells } from "@/components/cart-page/upsell-recommendations";
 import type { Locale } from "@/lib/i18n";
 import { getLocale } from "@/lib/params";
 import { getCart } from "@/lib/shopify/operations/cart";
